Add system::load_avg
Add function to return the system's load average. This includes the
average number of runnable entities averaged over one, five, and fifteen
minutes as well as statistics on the total number of runnable entities
and the last pid that was used.

Behind the scenes, the `try_parse` macro was moved to utils since it's
now used in multiple places.