Make all public types serde Serializable and Deserializable
Closes #83

The serde dependency needed to be renamed so that we can name a
feature with the same name: serde. We needed this because we needed
to pass down the serde feature to the platforms crate when it is
activated. From Cargo documentation, features cannot be named the
same as any dependency crate, you cannot have a feature named serde
if you depend on a crate also named serde. You can however rename
that crate, use the package option so that the right crate is still
fetched from crates.io, and then create your feature with the name.

Additionally, we needed to inform serde derive machinery of the new
crate name, and using `extern crate renamed_serde as serde` is not
sufficient and somehow the rename is not passed down. I could find
an issue about this: https://github.com/serde-rs/serde/issues/1855
it looks like the same problem, or close. To do that we needed to
add one additional serde attribute at every place derive is used.

All this ended up being quite complex and I think something has to
be done in Cargo, rustc or serde to make it easier.