Upgrade to nix 0.24.1, limit features
Limiting features removes memoffset as an indirect dependency, and should slightly decrease compile times.

nix 0.24 now correctly reports that uname() can fail, and that the returned fields are not guaranteed to be valid UTF-8.
Handle these by panicking if uname() fails (which should not happen), and reporting an empty string for invalid UTF-8.