Fix ProcessCpuTimes on macOS arm64 (Apple Silicon)
`proc_taskinfo` returns user and system time in mach tick units, not in
nanoseconds. On Intel machines, the mapping of mach tick units to
nanoseconds is 1/1. On Apple Silicon it is 3/125th. This constant can be
retrieved using `mach_timebase_info`.

https://developer.apple.com/documentation/driverkit/3433733-mach_timebase_info

A similar fix in htop: https://github.com/htop-dev/htop/pull/752