Update doc comments in Disk Usage
Spent a good bit of time to understand that this is the amount of disk free in disk. when looking at the api disk_usage(path), i wasn't sure if it had something to do with the path. Since 2 paths in the same mount will return the same result, it makes sense to update the comments to let folks know that this is independent of path